Remove legacy portal
